+++ cmd/ctfsmd/helpers.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"context"
-	"flag"
 	"fmt"
 	"io"
 	"log"
@@ -11,6 +10,7 @@ import (
 	"time"
 
 	"github.com/docker/docker/client"
+	"github.com/erodrigufer/CTForchestrator/internal/ctfsmd"
 	prometheus "github.com/erodrigufer/CTForchestrator/internal/prometheus"
 	semver "github.com/erodrigufer/go-semver"
 )
@@ -30,8 +30,9 @@ const charsetPassword = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ012
 // It configure all needed general parameters for the application, e.g. the
 // public port to which clients will connect.
 // Parameters: port, the port to which the clients will connect through SSH.
-func (app *application) setupApplication() error {
-	app.parseFlags()
+func (app *application) setupApplication(configValues ctfsmd.UserConfiguration) error {
+	// Fetch configValues
+	app.configurations = configValues
 
 	// Create a logger for INFO messages, the prefix "INFO" and a tab will be
 	// displayed before each log message. The flags Ldate and Ltime provide the
@@ -120,25 +121,6 @@ func (app *application) setupApplication() error {
 
 	return nil
 
-}
-
-// parseFlags, parses any flags if they are present.
-func (app *application) parseFlags() {
-	// Debug mode.
-	flag.BoolVar(&app.configurations.DebugMode, "debugMode", false, "Run daemon in 'debug' mode. Logging will be more extensive and frequent.")
-	// Parse ports options.
-	flag.StringVar(&app.configurations.SSHPort, "sshPort", "50000", "Port in which SSH Piper will work as an SSH proxy. Clients connect to this port.")
-	flag.StringVar(&app.configurations.HTTPAddr, "httpService", ":4000", "IP and port in which the HTTP webpage will be hosted.")
-	// Parse number of available and active sessions.
-	flag.IntVar(&app.configurations.MaxAvailableSess, "maxAvailableSess", 15, "Number of sessions always running in the background available to be delivered to clients.")
-	flag.IntVar(&app.configurations.MaxActiveSess, "maxActiveSess", 140, "Number of sessions that can be simultaneously actively being used by clients.")
-	// Parse lifetime of sessions and frequency to check for expired sessions.
-	flag.IntVar(&app.configurations.LifetimeSess, "lifetimeSess", 150, "Lifetime of session (in min) after which the session will expire.")
-	flag.IntVar(&app.configurations.SRDFreq, "srdFreq", 10, "Frequency (in min) with which srd checks if a session has expired.")
-	flag.IntVar(&app.configurations.TimeBetweenRequests, "timeBetweenRequests", 5, "Minimum time (in min) that has to pass between requests from the same IP so that request is not denied.")
-	// Debug mode.
-	flag.BoolVar(&app.configurations.NoInstrumentation, "noInstrumentation", false, "If true, no instrumentation will be performed by the application.")
-	flag.Parse()
 }
 
 // stopSession, stops all the containers that form a session and removes all