+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.1.0
+* Migrate CLI to `cobra` and `viper`.
+* Application can now be controlled by both flags and env. variables.
+
 ## v0.0.3
 * Remove old containers' references from the code's data model.
 * Refactor SSH Piper reverse proxy code to `reverseProxy.go`.